+++ src/ErrorHandler.php
@@ -4,7 +4,7 @@
  * Licensed under The MIT License
  * For full copyright and license information, please see the LICENSE file.
  * Redistributions of files must retain the above copyright notice.
- * 
+ *
  * @copyright (c) 2014, Achmad F. Ibrahim
  * @link https://github.com/Achsoft
  * @license http://opensource.org/licenses/mit-license.php The MIT License (MIT)
@@ -14,46 +14,56 @@ namespace Achsoft\Component\ErrorHandler;
 
 /**
  * Handle all Php errors as exceptions.
- * 
+ *
  * @author Achmad F. Ibrahim <acfatah@gmail.com>
  * @package Achsoft\Component\ErrorHandler
- * @version 0.1.1
+ * @version 0.2.0
  * @since 0.1.0
  */
 class ErrorHandler
 {
     /**
-     * Variable to store default exception class name.
-     * 
-     * @var string
-     * @since 0.1.0
-     */
-    private $exception = '\Achsoft\Component\ErrorHandler\Exception\ErrorException';
-    
-    /**
-     * Constructor.
-     * 
-     * @param string $class Exception class name
-     * @since 0.1.0
-     */
-    public function __construct($class = null)
-    {
-        is_null($class)
-            ? $this->setException($this->exception)
-            : $this->setException($class);
-    }
-    
-    /**
      * Exception error handler.
-     * 
-     * @since 0.1.0
+     *
+     * @since 0.2.0
      */
     public function handle($errno, $errstr, $errfile, $errline)
     {
-        $exception = $this->exception;
-        throw new $exception($errstr, 0, $errno, $errfile, $errline);
+         switch ($errno) {
+             case E_ERROR:
+                 throw new ErrorException($errstr, 0, $errno, $errfile, $errline);
+             case E_WARNING:
+                 throw new WarningException($errstr, 0, $errno, $errfile, $errline);
+             case E_PARSE:
+                 throw new ParseException($errstr, 0, $errno, $errfile, $errline);
+             case E_NOTICE:
+                 throw new NoticeException($errstr, 0, $errno, $errfile, $errline);
+             case E_CORE_ERROR:
+                 throw new CoreErrorException($errstr, 0, $errno, $errfile, $errline);
+             case E_CORE_WARNING:
+                 throw new CoreWarningException($errstr, 0, $errno, $errfile, $errline);
+             case E_COMPILE_ERROR:
+                 throw new CompileErrorException($errstr, 0, $errno, $errfile, $errline);
+             case E_COMPILE_WARNING:
+                 throw new CoreWarningException($errstr, 0, $errno, $errfile, $errline);
+             case E_USER_ERROR:
+                 throw new UserErrorException($errstr, 0, $errno, $errfile, $errline);
+             case E_USER_WARNING:
+                 throw new UserWarningException($errstr, 0, $errno, $errfile, $errline);
+             case E_USER_NOTICE:
+                 throw new UserNoticeException($errstr, 0, $errno, $errfile, $errline);
+             case E_STRICT:
+                 throw new StrictException($errstr, 0, $errno, $errfile, $errline);
+             case E_RECOVERABLE_ERROR:
+                 throw new RecoverableErrorException($errstr, 0, $errno, $errfile, $errline);
+             case E_DEPRECATED:
+                 throw new DeprecatedException($errstr, 0, $errno, $errfile, $errline);
+             case E_USER_DEPRECATED:
+                 throw new UserDeprecatedException($errstr, 0, $errno, $errfile, $errline);
+        }
+
     }
-    
+
     /**
      * Register exception handler.
      *
@@ -66,7 +76,7 @@ class ErrorHandler
         ini_set('html_errors', false);
         set_error_handler([$this, 'handle']);
     }
-    
+
     /**
      * Restore default error handler.
      *
@@ -75,16 +85,5 @@ class ErrorHandler
     public function restore()
     {
         restore_error_handler();
-    }
-    
-    /**
-     * Set exception class.
-     * 
-     * @param string $class Fully qualified exception class name
-     * @since 0.1.0
-     */
-    public function setException($class)
-    {
-        $this->exception = $class;
     }
 }