+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0 (2014.07.08)
+* Add exception types based on E_* error types
+* Remove constructor
+* Remove default error exception and its setter
+
 ## 0.1.1 (2014.07.05)
 * Add a restore method to restore default error handler